+++ NEWS.md
@@ -11,6 +11,7 @@
 * Fix self-induced preemption through `set_capacity()` (#125).
 * Update "Queueing Systems" vignette (a0409a0, 8f03f4f).
 * Update "Advanced Trajectory Usage" vignette (4501927).
+* Fix print methods to return the object invisibly (#128).
 
 # simmer 3.6.4